requirements.txt
# Core dependencies for smart document classification
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0

# Optional but recommended for better PDF processing
# pdfplumber>=0.9.0  # Better text and table extraction
# tabula-py>=2.7.0   # Advanced table extraction